Improve responsivity for TaskGroups
