+++ src/pages/Board.jsx
@@ -1,15 +1,30 @@
+/* eslint-disable no-console */
 import * as React from 'react'
 import { useParams } from 'react-router-dom'
 import { Box, Grid, Button, Text } from '@chakra-ui/react'
-import { getBoard } from '../utils/api'
-
-import TaskGrpups from './components/TaskGroups'
+import { PlusSquareIcon } from '@chakra-ui/icons'
+import { getBoard, getTaskGroups } from '../utils/api'
+import Task from './components/Tasks'
 
 const Board = () => {
   const { id } = useParams()
+
   // eslint-disable-next-line no-unused-vars
   const [status, setStatus] = React.useState('loading')
   const [board, setBoard] = React.useState({})
+  const [groups, setGroups] = React.useState([])
+
+  React.useEffect(() => {
+    try {
+      const fetchData = async () => {
+        const data = await getTaskGroups(id)
+        setGroups(data)
+      }
+      fetchData()
+    } catch (e) {
+      console.log(e)
+    }
+  }, [id])
 
   React.useEffect(() => {
     try {
@@ -26,14 +41,63 @@ const Board = () => {
 
   return (
     <Box h="100vh" bg="gray.100" p="2">
-      <Text p="3" fontWeight="bold">
+      <Text textAlign="center" fontSize="xx-large" p="3" fontWeight="bold">
         {board.name}
       </Text>
-
-      <Grid mt="3" templateColumns={{ base: '1fr 1fr', lg: '1fr 1fr 1fr 1fr' }}>
-        <TaskGrpups key boardId={board.id} />
-
-        <Button colorScheme="blue">+ Add new group </Button>
+      <Grid gridTemplateColumns={{ base: '1fr 1fr', lg: '1fr 1fr 1fr 1fr' }}>
+        {groups.map((group) => (
+          <Box
+            key={group.id}
+            m="2"
+            borderWidth="1px"
+            borderRadius="lg"
+            overflow="hidden"
+            backgroundColor="#F7FAFC"
+            s
+          >
+            <Box
+              m="2"
+              backgroundColor="blue.400"
+              textColor="white"
+              borderRadius="lg"
+              textTransform="uppercase"
+              letterSpacing="wide"
+              fontWeight="bold"
+              textAlign="center"
+              p="4"
+            >
+              {group.name}
+            </Box>
+            <Task boardId={id} />
+            <Box m="2">
+              <Button w="100%" color="white" leftIcon={<PlusSquareIcon />} background="blue.400">
+                Add new task
+              </Button>
+            </Box>
+          </Box>
+        ))}
+        <Box
+          textColor="white"
+          borderRadius="lg"
+          textTransform="uppercase"
+          letterSpacing="wide"
+          fontWeight="bold"
+          textAlign="center"
+          p="4"
+        >
+          <Button
+            textTransform="uppercase"
+            letterSpacing="wide"
+            fontWeight="bold"
+            color="white"
+            leftIcon={<PlusSquareIcon />}
+            p="7"
+            w="100%"
+            background="blue.400"
+          >
+            Add new group
+          </Button>
+        </Box>
       </Grid>
     </Box>
   )
